[QFEED-141] Refactor: socialnetwork 컨텍스트 디렉토리 재구성
