+++ module-application/src/main/java/com/wsws/moduleapplication/feed/service/AnswerReadService.java
@@ -10,8 +10,8 @@ import com.wsws.moduledomain.feed.answer.repo.AnswerRepository;
 import com.wsws.moduledomain.feed.comment.AnswerComment;
 import com.wsws.moduledomain.feed.comment.repo.AnswerCommentRepository;
 import com.wsws.moduledomain.feed.dto.AnswerQuestionDTO;
-import com.wsws.moduledomain.follow.aggregate.Follow;
-import com.wsws.moduledomain.follow.repo.FollowRepository;
+import com.wsws.moduledomain.socialnetwork.follow.aggregate.Follow;
+import com.wsws.moduledomain.socialnetwork.follow.repo.FollowRepository;
 import com.wsws.moduledomain.feed.like.Like;
 import com.wsws.moduledomain.usercontext.user.aggregate.User;
 import com.wsws.moduledomain.feed.like.LikeRepository;
