+++ module-infra/src/main/java/com/wsws/moduleinfra/socialnetworkcontext/follow/repository/FollowRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.wsws.moduleinfra.socialnetworkcontext.follow.repository;
 
 
-import com.wsws.moduledomain.follow.aggregate.Follow;
-import com.wsws.moduledomain.follow.repo.FollowRepository;
+import com.wsws.moduledomain.socialnetwork.follow.aggregate.Follow;
+import com.wsws.moduledomain.socialnetwork.follow.repo.FollowRepository;
 import com.wsws.moduleinfra.socialnetworkcontext.follow.entity.FollowEntity;
 import com.wsws.moduleinfra.socialnetworkcontext.follow.entity.FollowEntityMapper;
 import lombok.RequiredArgsConstructor;
