+++ module-infra/src/main/java/com/wsws/moduleinfra/socialnetworkcontext/interest/repo/UserInterestRepositoryImpl.java
@@ -5,11 +5,11 @@ import com.wsws.moduledomain.socialnetwork.interest.UserInterestRepository;
 import com.wsws.moduledomain.usercontext.user.vo.UserId;
 import com.wsws.moduledomain.socialnetwork.interest.UserInterest;
 import com.wsws.moduleinfra.entity.CategoryEntity;
-import com.wsws.moduleinfra.usercontext.entity.UserEntity;
+import com.wsws.moduleinfra.usercontext.user.entity.UserEntity;
 import com.wsws.moduleinfra.socialnetworkcontext.interest.entity.UserInterestEntity;
 import com.wsws.moduleinfra.socialnetworkcontext.interest.entity.UserInterestEntityMapper;
 import com.wsws.moduleinfra.repo.category.CategoryJpaRepository;
-import com.wsws.moduleinfra.usercontext.user.JpaUserRepository;
+import com.wsws.moduleinfra.usercontext.user.repo.JpaUserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
